# clvImob-main/server/swagger.json
{
    "openapi" : "3.0.0",
    "info" : {
        "Title": "CLVImob API",
        "description": "Consult API end-points to the site",
        "contact": "cheljose@gmail.com",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000/",
            "description": "Local server"
        }
    ],
    "paths": {
...
    "/pagamentos": {
        "post": {
            "summary": "Cria um novo pagamento",
            "security": [{"authMiddleware": []}],
            "operationId": "createPagamento",
            "responses": {"200": {"description": "Pagamento criado com sucesso"}}
        },
        "get": {
            "summary": "Obtém todos os pagamentos",
            "security": [{"authMiddleware": []}],
            "operationId": "getAllPagamentos",
            "responses": {"200": {"description": "Pagamentos obtidos com sucesso"}}
        }
    },
    "/pagamentos/{id}": {
        "get": {
            "summary": "Obtém um pagamento específico por ID",
            "security": [{"authMiddleware": []}],
            "operationId": "getPagamentoById",
            "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
            "responses": {"200": {"description": "Pagamento obtido com sucesso"}}
        },
        "delete": {
            "summary": "Deleta um pagamento específico por ID",
            "security": [{"authMiddleware": []}],
            "operationId": "deletePagamento",
            "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
            "responses": {"200": {"description": "Pagamento deletado com sucesso"}}
        },
        "put": {
            "summary": "Atualiza um pagamento específico por ID",
            "security": [{"authMiddleware": []}],
            "operationId": "updatePagamento",
            "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
            "responses": {"200": {"description": "Pagamento atualizado com sucesso"}}
        }
    },
...
    },
...
}